+++ launcher/__init__.py
@@ -1,0 +1,1 @@
+"""Bear Cave PTR launcher."""